+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [unReleased] - 2022-07-16
+
+### changed
+
+- Refactor container Deque from function to class (don't effect using)
+
 ## [3.0.0-beta.0] - 2022-04-29
 
 ### Added